requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
passlib==1.7.4
bcrypt==4.0.1
python-multipart
python-jose[cryptography]